Change displayExchangeUI
This is being made to make more sense upon what we are doing, we are currently displaying the table elements in the UI not the whole UI as one.
