+++ src/ui/utils.js
@@ -1,6 +1,6 @@
 import { handleInputCurrency, handleInputDate, handleListChange } from "./handlers.js";
 import { displayLoadingTable, clearExchangeTable } from "./table.js";
-import { displayExchangeUI } from "./exchange.js";
+import { displayExchangeTable } from "./exchange.js";
 
 export function setupListChanges(list, date) {
   list.onclick = (currency) => {
@@ -9,7 +9,7 @@ export function setupListChanges(list, date) {
 
     clearExchangeTable();
     displayLoadingTable();
-    displayExchangeUI(clickedCurrency, selectedDate);
+    displayExchangeTable(clickedCurrency, selectedDate);
   };
 }
 
@@ -20,7 +20,7 @@ export function setupCurrencyChanges(currency, date) {
 
     clearExchangeTable();
     displayLoadingTable();
-    displayExchangeUI(selectedCurrency, selectedDate);
+    displayExchangeTable(selectedCurrency, selectedDate);
   };
 }
 
@@ -32,35 +32,9 @@ export function setupDateChanges(date, currency) {
     setTimeout(() => {
       const selectedCurrency = handleInputCurrency(currency);
       const selectedDate = handleInputDate(date);
-      displayExchangeUI(selectedCurrency, selectedDate);
+      displayExchangeTable(selectedCurrency, selectedDate);
     }, 1000);
   };
-}
-
-export function setCurrencyTitle(base, currency) {
-  const baseCurrency = base;
-  const currentInputValue = currency;
-  const currentTitle = document.querySelector("#current-currency");
-  const $code = document.querySelectorAll(".currency-code");
-
-  for (const code of $code) {
-    if (currentInputValue === code.innerText) {
-      return (currentTitle.innerText = `Currently displaying ${currentInputValue}`);
-    }
-  }
-  return (currentTitle.innerText = `Currently displaying ${baseCurrency}`);
-}
-
-export function setDateTitle(present, date) {
-  const presentDate = present;
-  const currentInputDate = date;
-  const currentDateTitle = document.querySelector("#current-date");
-
-  if (currentInputDate && currentInputDate !== "latest") {
-    return (currentDateTitle.innerText = `At ${currentInputDate} as date of exchange`);
-  }
-
-  return (currentDateTitle.innerText = `At ${presentDate} as date of exchange`);
 }
 
 export function getCurrencyCode(currency) {
